perf(frontend): split LiveKit/hls out of the initial bundle
Entry chunk shrinks from 2.95MB to 1.28MB raw (801KB to 327KB gzip) by
moving livekit-client, hls.js, and MUI into their own manualChunks, and
by dynamic-importing livekit-client at the actual join seam instead of
loading it eagerly for every user.

- voiceActions.ts: Room/RoomEvent/DisconnectReason/soundboardPlayer are
  now dynamically imported inside connectToLiveKitRoom() and the
  soundboard call sites — the only places that construct/use a Room.
  This happens inside the existing "Connecting…" state, so the join
  spinner already covers the chunk-load latency.
- RoomContext/RoomContextDef and ~10 other hook/component files:
  convert livekit-client symbols to `import type` wherever they're only
  used as TS types, leaving real runtime imports (enums like Track,
  RoomEvent) only where genuinely needed.
- VoiceBottomBar split into a thin, always-mounted shell (no livekit
  import) + VoiceBottomBarContent, lazy-loaded only once connected.
  VideoTiles, VoiceDebugPanel, and TrimPreview (the sole hls.js
  importer) are React.lazy at their mount points.
- vite.config.ts: manualChunks buckets livekit-client/@livekit/
  components-react, hls.js, and MUI into separate cacheable chunks;
  precache limit tightened from 6MB to 3MB now that no chunk is near
  that size.
- rollup-plugin-visualizer added as a dev-only, ANALYZE=true-gated
  plugin for bundle measurement (`pnpm run build:analyze`).

Remaining eager livekit-client imports (AudioRenderer,
TrackSubscriptionProvider, VoiceEventLogProvider) are always-mounted in
Layout.tsx and need live runtime enums (Track, RoomEvent) as event
names — manualChunks keeps them out of the entry chunk, but they still
trigger a <link rel="modulepreload"> for the livekit chunk on every
authenticated page load. See the PR-11 report for the full
categorization and the case for a follow-up to gate those three behind
the connected state.
