+++ frontend/src/components/Voice/VoiceBottomBar.tsx
@@ -1,647 +1,40 @@
-import React, { useState, useCallback, useRef } from "react";
-import {
-  Box,
-  Paper,
-  Typography,
-  IconButton,
-  Tooltip,
-  Chip,
-  Divider,
-  Menu,
-  MenuItem,
-  Badge,
-} from "@mui/material";
-import { useTheme } from "@mui/material/styles";
-import {
-  Mic,
-  MicOff,
-  Headset,
-  HeadsetOff,
-  Videocam,
-  VideocamOff,
-  ScreenShare,
-  StopScreenShare,
-  CallEnd,
-  Settings,
-  VolumeUp,
-  FiberManualRecord,
-  MovieCreation,
-  VideoCall,
-  SpeakerPhone,
-  PhoneInTalk,
-} from "@mui/icons-material";
-import { useNavigate } from "react-router-dom";
+import React, { Suspense, lazy, useEffect } from "react";
 import { useVoiceConnection } from "../../hooks/useVoiceConnection";
-import { useScreenShare } from "../../hooks/useScreenShare";
-import { useLocalMediaState } from "../../hooks/useLocalMediaState";
-import { useDeafenEffect } from "../../hooks/useDeafenEffect";
-import { useReplayBufferState } from "../../contexts/ReplayBufferContext";
-import { useDebugPanelShortcut } from "../../hooks/useDebugPanelShortcut";
-import { usePushToTalk } from "../../hooks/usePushToTalk";
-import { DeviceSettingsDialog } from "./DeviceSettingsDialog";
-import { SoundboardButton } from "./SoundboardButton";
-import { ScreenSourcePicker } from "./ScreenSourcePicker";
-import { VoiceDebugPanel } from "./VoiceDebugPanel";
-import { CaptureReplayModal } from "./CaptureReplayModal";
-import { useResponsive } from "../../hooks/useResponsive";
-import { useHapticFeedback } from "../../hooks/useHapticFeedback";
-import { useWakeLock } from "../../hooks/useWakeLock";
-import { logger } from "../../utils/logger";
-import { LAYOUT_CONSTANTS } from "../../utils/breakpoints";
-import { useSpeaking } from "../../hooks/useSpeaking";
-import { useVoicePresenceHeartbeat } from "../../hooks/useVoicePresenceHeartbeat";
-import { useBackgroundVoiceKeepAlive } from "../../hooks/useBackgroundVoiceKeepAlive";
-import { useVoiceMediaSession } from "../../hooks/useVoiceMediaSession";
-import { useVoiceForegroundResync } from "../../hooks/useVoiceForegroundResync";
-import { useServerMuteEffect } from "../../hooks/useServerMuteEffect";
-import { useRemoteVolumeEffect } from "../../hooks/useRemoteVolumeEffect";
-import { useCurrentUser } from "../../hooks/useCurrentUser";
-import { VoiceSessionType } from "../../contexts/VoiceContext";
 
+// The real bottom-bar UI (VoiceBottomBarContent) pulls in every voice-session
+// hook — including several that statically import livekit-client for runtime
+// enums (Track, RoomEvent, ...). Loading it lazily keeps that whole subtree,
+// and the LiveKit chunk it needs, out of the always-mounted app shell for
+// users who never join voice. See PR-11 (bundle splitting).
+const VoiceBottomBarContent = lazy(() => import("./VoiceBottomBarContent"));
+
+/**
+ * Always-mounted shell (see Layout.tsx). Renders nothing until voice is
+ * connected, then Suspense-loads the real bar. Kept intentionally thin and
+ * free of any livekit-client import so it never contributes to the eager
+ * bundle graph.
+ */
 export const VoiceBottomBar: React.FC = () => {
-  const theme = useTheme();
-  const navigate = useNavigate();
-  const { state, actions } = useVoiceConnection();
-  const screenShare = useScreenShare();
-  const { isCameraEnabled, isMicrophoneEnabled } = useLocalMediaState();
-  const { isMobile, shouldUseTouchUI } = useResponsive();
-  const haptic = useHapticFeedback();
-  const { user: currentUser } = useCurrentUser();
-  const { isSpeaking } = useSpeaking();
-  const [settingsAnchor, setSettingsAnchor] = useState<null | HTMLElement>(
-    null
-  );
-  const [showDeviceSettings, setShowDeviceSettings] = useState(false);
-  const [showCaptureModal, setShowCaptureModal] = useState(false);
-  const [isSpeakerphone, setIsSpeakerphone] = useState(false);
+  const { state } = useVoiceConnection();
+  const isActive = state.isConnected && (!!state.currentChannelId || !!state.currentDmGroupId);
 
-  // Use extracted hooks for cleaner organization
-  const { showDebugPanel } = useDebugPanelShortcut();
-  const {
-    isActive: isPTTActive,
-    isKeyHeld: isPTTKeyHeld,
-    currentKeyDisplay: pttKeyDisplay,
-    pttPress,
-    pttRelease,
-  } = usePushToTalk();
-
-  // On touch devices the PTT key can't be pressed, so the mic button becomes a
-  // hold-to-talk control. Desktop PTT (keyboard) and non-PTT tap-to-mute are
-  // unaffected.
-  const isHoldToTalk = isPTTActive && shouldUseTouchUI;
-
-  // Prevent tab freeze / OS suspension while in voice
-  useBackgroundVoiceKeepAlive({ isConnected: state.isConnected });
-
-  // Keep the screen awake while connected (touch devices / web); no-op in
-  // Electron and where the Wake Lock API is unsupported.
-  useWakeLock(state.isConnected);
-
-  // Surface the call via the Media Session API (Android ongoing-call
-  // notification + higher background process priority, see #350)
-  useVoiceMediaSession({
-    isConnected: state.isConnected,
-    contextName: state.channelName ?? state.dmGroupName,
-    isMicrophoneEnabled,
-    onHangup: actions.leaveVoiceChannel,
-    onToggleMic: actions.toggleMute,
-  });
-
-  // Recover calls that silently died while backgrounded/locked (#350)
-  useVoiceForegroundResync({
-    room: state.room,
-    state,
-    actions: {
-      joinVoiceChannel: actions.joinVoiceChannel,
-      joinDmVoice: actions.joinDmVoice,
-    },
-  });
-
-  // Keep voice presence TTL alive in Redis while connected
-  useVoicePresenceHeartbeat({
-    channelId: state.currentChannelId,
-    dmGroupId: state.currentDmGroupId,
-    contextType: state.contextType,
-  });
-
-  // Implement proper deafen functionality (mute received audio)
-  useDeafenEffect();
-
-  // Listen for server mute WS events and enforce on local user
-  useServerMuteEffect();
-
-  // Reapply per-user volume from localStorage when remote tracks change
-  useRemoteVolumeEffect();
-
-  // Automatically manage replay buffer when screen sharing
-  const { isReplayBufferActive } = useReplayBufferState();
-
-  // Check if the current user is speaking
-  const isCurrentUserSpeaking = currentUser ? isSpeaking(currentUser.id) : false;
-
-  // Define callbacks before any early returns (React hooks must be called unconditionally)
-  const handleSettingsClick = useCallback((event: React.MouseEvent<HTMLElement>) => {
-    setSettingsAnchor(event.currentTarget);
-  }, []);
-
-  const handleSettingsClose = useCallback(() => {
-    setSettingsAnchor(null);
-  }, []);
-
-  const handleToggleVideo = useCallback(() => {
-    actions.toggleVideo();
-  }, [actions]);
-
-  const handleDeviceSettingsOpen = useCallback(() => {
-    setShowDeviceSettings(true);
-    setSettingsAnchor(null);
-  }, []);
-
-  const handleDeviceSettingsClose = useCallback(() => {
-    setShowDeviceSettings(false);
-  }, []);
-
-  const handleDeviceChange = useCallback(async (type: 'audio' | 'video' | 'audioOutput', deviceId: string) => {
-    try {
-      if (type === 'audio') {
-        await actions.switchAudioInputDevice(deviceId);
-      } else if (type === 'audioOutput') {
-        await actions.switchAudioOutputDevice(deviceId);
-      } else if (type === 'video') {
-        await actions.switchVideoInputDevice(deviceId);
-      }
-    } catch (error) {
-      logger.error(`Failed to switch ${type} device:`, error);
+  // Warm the content chunk as soon as a join starts (the existing
+  // "Connecting…" phase), so it's already cached by the time isConnected
+  // flips true and this Suspense boundary would otherwise show a gap after
+  // the LiveKit chunk itself has already loaded via the join flow.
+  useEffect(() => {
+    if (state.isConnecting) {
+      void import("./VoiceBottomBarContent");
     }
-  }, [actions]);
+  }, [state.isConnecting]);
 
-  const handleToggleScreenShare = useCallback(() => {
-    screenShare.toggleScreenShare();
-  }, [screenShare]);
-
-  // Hold-to-talk (touch): press engages the mic through the shared PTT logic
-  // (which enforces the server-mute guard), release re-mutes. Pointer capture
-  // keeps transmit tied to this button even if the finger slides off it, so
-  // release is driven by pointerup / pointercancel / lostpointercapture — not
-  // pointerleave, which would end the hold mid-slide. holdEngagedRef makes the
-  // release path idempotent across those overlapping events.
-  const holdEngagedRef = useRef(false);
-  const handlePttPointerDown = useCallback((e: React.PointerEvent<HTMLButtonElement>) => {
-    if (state.isServerMuted) return;
-    e.currentTarget.setPointerCapture?.(e.pointerId);
-    holdEngagedRef.current = true;
-    haptic.medium();
-    void pttPress?.();
-  }, [state.isServerMuted, haptic, pttPress]);
-
-  const handlePttPointerUp = useCallback(() => {
-    if (!holdEngagedRef.current) return;
-    holdEngagedRef.current = false;
-    haptic.light();
-    void pttRelease?.();
-  }, [haptic, pttRelease]);
-
-  // Check if browser supports audio output switching (setSinkId)
-  const supportsSpeakerToggle = isMobile && 'setSinkId' in HTMLMediaElement.prototype;
-
-  const handleToggleSpeakerphone = useCallback(async () => {
-    try {
-      const newSpeakerState = !isSpeakerphone;
-
-      // Enumerate real audio output devices to find valid IDs
-      const devices = await navigator.mediaDevices.enumerateDevices();
-      const audioOutputs = devices.filter(d => d.kind === 'audiooutput');
-
-      // "communications" is the earpiece on some platforms; "default" is the system default (speaker on mobile)
-      const earpieceDevice = audioOutputs.find(d => d.deviceId === 'communications');
-      const defaultDevice = audioOutputs.find(d => d.deviceId === 'default') ?? audioOutputs[0];
-
-      const targetDeviceId = newSpeakerState
-        ? (defaultDevice?.deviceId ?? 'default')
-        : (earpieceDevice?.deviceId ?? 'default');
-
-      await actions.switchAudioOutputDevice(targetDeviceId);
-      setIsSpeakerphone(newSpeakerState);
-    } catch (error) {
-      logger.error('Failed to toggle speakerphone:', error);
-    }
-  }, [isSpeakerphone, actions]);
-
-  // Show bar if connected to either a channel or DM
-  if (!state.isConnected || (!state.currentChannelId && !state.currentDmGroupId)) {
+  if (!isActive) {
     return null;
   }
 
-  // Determine display name and type
-  const displayName = state.contextType === VoiceSessionType.Dm
-    ? state.dmGroupName || 'Direct Message'
-    : state.channelName || 'Voice Channel';
-
-  const displayType = state.contextType === VoiceSessionType.Dm ? 'DM Voice Call' : 'Voice Connected';
-
   return (
-    <>
-      {/* Main Bottom Bar */}
-      <Paper
-        elevation={8}
-        sx={{
-          position: "fixed",
-          bottom: 0,
-          left: 0,
-          right: 0,
-          zIndex: 1300,
-          borderRadius: 0,
-          backgroundColor: "background.paper",
-          borderTop: 1,
-          borderColor: "divider",
-        }}
-      >
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "space-between",
-            px: isMobile ? 1 : 3,
-            py: isMobile ? 1 : 1.5,
-            minHeight: isMobile ? LAYOUT_CONSTANTS.VOICE_BAR_HEIGHT_MOBILE : 64,
-            gap: isMobile ? 0.5 : 1,
-          }}
-        >
-          {/* Channel/DM Info */}
-          <Box sx={{ display: "flex", alignItems: "center", gap: isMobile ? 0.5 : 2, flex: 1, minWidth: 0 }}>
-            <Box sx={{ display: "flex", alignItems: "center", gap: 1, minWidth: 0 }}>
-              <VolumeUp color="primary" sx={{ flexShrink: 0 }} />
-              <Box sx={{ minWidth: 0 }}>
-                <Typography variant="body2" fontWeight="medium" noWrap>
-                  {displayName}
-                </Typography>
-                {!isMobile && (
-                  <Typography variant="caption" color="text.secondary">
-                    {displayType}
-                  </Typography>
-                )}
-              </Box>
-            </Box>
-
-            {/* Connection Status - hide on mobile */}
-            {!isMobile && (
-              <Chip
-                label={state.isConnected ? "Connected" : "Connecting..."}
-                color={state.isConnected ? "success" : "warning"}
-                size="small"
-                sx={{ height: 24 }}
-              />
-            )}
-
-          </Box>
-
-          {/* Voice Controls */}
-          <Box sx={{ display: "flex", alignItems: "center", gap: isMobile ? 0.5 : 1 }}>
-            {/* Microphone */}
-            <Tooltip
-              title={
-                state.isServerMuted
-                  ? "Server Muted — contact a moderator"
-                  : isHoldToTalk
-                    ? (isPTTKeyHeld ? "Transmitting..." : "Hold to talk")
-                    : isPTTActive
-                      ? (isPTTKeyHeld ? "Transmitting..." : `Hold ${pttKeyDisplay} to talk`)
-                      : (!isMicrophoneEnabled ? "Unmute" : "Mute")
-              }
-              arrow={!isMobile}
-              disableTouchListener={isHoldToTalk}
-            >
-              <IconButton
-                onClick={isPTTActive || state.isServerMuted ? undefined : actions.toggleMute}
-                onPointerDown={isHoldToTalk ? handlePttPointerDown : undefined}
-                onPointerUp={isHoldToTalk ? handlePttPointerUp : undefined}
-                onPointerCancel={isHoldToTalk ? handlePttPointerUp : undefined}
-                onLostPointerCapture={isHoldToTalk ? handlePttPointerUp : undefined}
-                onContextMenu={isHoldToTalk ? (e) => e.preventDefault() : undefined}
-                color={!isMicrophoneEnabled && !isPTTKeyHeld ? "error" : "default"}
-                size={isMobile ? "medium" : "medium"}
-                sx={{
-                  touchAction: isHoldToTalk ? "none" : undefined,
-                  userSelect: isHoldToTalk ? "none" : undefined,
-                  WebkitUserSelect: isHoldToTalk ? "none" : undefined,
-                  backgroundColor: isPTTKeyHeld
-                    ? theme.palette.semantic.status.positive
-                    : state.isServerMuted
-                      ? "warning.main"
-                      : (!isMicrophoneEnabled ? "error.main" : "transparent"),
-                  color: isPTTKeyHeld
-                    ? theme.palette.getContrastText(theme.palette.semantic.status.positive)
-                    : state.isServerMuted
-                      ? "warning.contrastText"
-                      : (!isMicrophoneEnabled ? "error.contrastText" : "text.primary"),
-                  minWidth: isMobile ? 48 : "auto",
-                  minHeight: isMobile ? 48 : "auto",
-                  border: (isMicrophoneEnabled && isCurrentUserSpeaking) || isPTTKeyHeld
-                    ? `2px solid ${theme.palette.semantic.status.positive}`
-                    : "2px solid transparent",
-                  boxShadow: (isMicrophoneEnabled && isCurrentUserSpeaking) || isPTTKeyHeld
-                    ? `0 0 8px ${theme.palette.semantic.status.positive}`
-                    : "none",
-                  transition: "all 0.2s ease",
-                  cursor: isPTTActive || state.isServerMuted ? "default" : "pointer",
-                  "&:hover": {
-                    backgroundColor: isPTTKeyHeld
-                      ? theme.palette.semantic.status.positive
-                      : state.isServerMuted
-                        ? "warning.dark"
-                        : (!isMicrophoneEnabled
-                          ? "error.dark"
-                          : "action.hover"),
-                  },
-                }}
-              >
-                {!isMicrophoneEnabled && !isPTTKeyHeld ? <MicOff /> : <Mic />}
-              </IconButton>
-            </Tooltip>
-
-            {/* Headphones/Deafen - hide on mobile */}
-            {!isMobile && (
-              <Tooltip title={state.isDeafened ? "Undeafen" : "Deafen"}>
-                <IconButton
-                  onClick={actions.toggleDeafen}
-                  color={state.isDeafened ? "error" : "default"}
-                  sx={{
-                    backgroundColor: state.isDeafened
-                      ? "error.main"
-                      : "transparent",
-                    color: state.isDeafened
-                      ? "error.contrastText"
-                      : "text.primary",
-                    "&:hover": {
-                      backgroundColor: state.isDeafened
-                        ? "error.dark"
-                        : "action.hover",
-                    },
-                  }}
-                >
-                  {state.isDeafened ? <HeadsetOff /> : <Headset />}
-                </IconButton>
-              </Tooltip>
-            )}
-
-            {/* Speakerphone toggle - mobile only, when browser supports setSinkId */}
-            {supportsSpeakerToggle && (
-              <Tooltip title={isSpeakerphone ? "Switch to earpiece" : "Switch to speaker"}>
-                <IconButton
-                  onClick={handleToggleSpeakerphone}
-                  size="medium"
-                  sx={{
-                    backgroundColor: isSpeakerphone
-                      ? "primary.main"
-                      : "transparent",
-                    color: isSpeakerphone
-                      ? "primary.contrastText"
-                      : "text.primary",
-                    minWidth: 48,
-                    minHeight: 48,
-                    "&:hover": {
-                      backgroundColor: isSpeakerphone
-                        ? "primary.dark"
-                        : "action.hover",
-                    },
-                  }}
-                >
-                  {isSpeakerphone ? <SpeakerPhone /> : <PhoneInTalk />}
-                </IconButton>
-              </Tooltip>
-            )}
-
-            <Divider orientation="vertical" flexItem sx={{ mx: isMobile ? 0.5 : 1 }} />
-
-            {/* Video */}
-            <Tooltip
-              title={
-                isCameraEnabled ? "Turn off camera" : "Turn on camera"
-              }
-              arrow={!isMobile}
-            >
-              <IconButton
-                onClick={handleToggleVideo}
-                color={isCameraEnabled ? "primary" : "default"}
-                size={isMobile ? "medium" : "medium"}
-                sx={{
-                  backgroundColor: isCameraEnabled
-                    ? "primary.main"
-                    : "transparent",
-                  color: isCameraEnabled
-                    ? "primary.contrastText"
-                    : "text.primary",
-                  minWidth: isMobile ? 48 : "auto",
-                  minHeight: isMobile ? 48 : "auto",
-                  "&:hover": {
-                    backgroundColor: isCameraEnabled
-                      ? "primary.dark"
-                      : "action.hover",
-                  },
-                }}
-              >
-                {isCameraEnabled ? <Videocam /> : <VideocamOff />}
-              </IconButton>
-            </Tooltip>
-
-            {/* Screen Share */}
-            <Tooltip
-              title={
-                screenShare.isScreenSharing ? "Stop screen share" : "Share screen"
-              }
-              arrow={!isMobile}
-            >
-              <Badge
-                overlap="circular"
-                anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-                badgeContent={
-                  isReplayBufferActive ? (
-                    <FiberManualRecord
-                      sx={{
-                        width: 8,
-                        height: 8,
-                        color: theme.palette.semantic.status.positive,
-                        animation: 'pulse 1.5s ease-in-out infinite',
-                        '@keyframes pulse': {
-                          '0%, 100%': { opacity: 1 },
-                          '50%': { opacity: 0.5 },
-                        },
-                      }}
-                    />
-                  ) : null
-                }
-              >
-                <IconButton
-                  onClick={handleToggleScreenShare}
-                  color={screenShare.isScreenSharing ? "primary" : "default"}
-                  size={isMobile ? "medium" : "medium"}
-                  sx={{
-                    backgroundColor: screenShare.isScreenSharing
-                      ? "primary.main"
-                      : "transparent",
-                    color: screenShare.isScreenSharing
-                      ? "primary.contrastText"
-                      : "text.primary",
-                    minWidth: isMobile ? 48 : "auto",
-                    minHeight: isMobile ? 48 : "auto",
-                    "&:hover": {
-                      backgroundColor: screenShare.isScreenSharing
-                        ? "primary.dark"
-                        : "action.hover",
-                    },
-                  }}
-                >
-                  {screenShare.isScreenSharing ? <StopScreenShare /> : <ScreenShare />}
-                </IconButton>
-              </Badge>
-            </Tooltip>
-
-            {/* Capture Replay - only show when replay buffer is active */}
-            {isReplayBufferActive && (
-              <Tooltip
-                title="Capture Replay"
-                arrow={!isMobile}
-              >
-                <IconButton
-                  onClick={() => setShowCaptureModal(true)}
-                  color="success"
-                  size={isMobile ? "medium" : "medium"}
-                  sx={{
-                    minWidth: isMobile ? 48 : "auto",
-                    minHeight: isMobile ? 48 : "auto",
-                    "&:hover": {
-                      backgroundColor: "success.main",
-                      color: "success.contrastText",
-                    },
-                  }}
-                >
-                  <MovieCreation />
-                </IconButton>
-              </Tooltip>
-            )}
-
-            {/* Soundboard - community voice channels only */}
-            {state.isConnected && state.communityId && actions.playSoundboard && (
-              <SoundboardButton
-                communityId={state.communityId}
-                onPlay={actions.playSoundboard}
-                isMobile={isMobile}
-              />
-            )}
-
-            {/* Show Video Tiles - visible when tiles are hidden and user is connected */}
-            {!state.showVideoTiles && state.isConnected && (
-              <Tooltip title="Show Video Tiles" arrow={!isMobile}>
-                <IconButton
-                  onClick={() => actions.setShowVideoTiles(true)}
-                  size={isMobile ? "medium" : "medium"}
-                  sx={{
-                    minWidth: isMobile ? 48 : "auto",
-                    minHeight: isMobile ? 48 : "auto",
-                    "&:hover": {
-                      backgroundColor: "action.hover",
-                    },
-                  }}
-                >
-                  <VideoCall />
-                </IconButton>
-              </Tooltip>
-            )}
-
-            {/* Settings - hide on mobile, use menu instead */}
-            {!isMobile && (
-              <>
-                <Divider orientation="vertical" flexItem sx={{ mx: 1 }} />
-                <Tooltip title="Voice settings">
-                  <IconButton onClick={handleSettingsClick}>
-                    <Settings />
-                  </IconButton>
-                </Tooltip>
-              </>
-            )}
-
-            <Divider orientation="vertical" flexItem sx={{ mx: isMobile ? 0.5 : 1 }} />
-
-            {/* Disconnect */}
-            <Tooltip title="Disconnect" arrow={!isMobile}>
-              <IconButton
-                onClick={actions.leaveVoiceChannel}
-                color="error"
-                size={isMobile ? "medium" : "medium"}
-                sx={{
-                  minWidth: isMobile ? 48 : "auto",
-                  minHeight: isMobile ? 48 : "auto",
-                  "&:hover": {
-                    backgroundColor: "error.main",
-                    color: "error.contrastText",
-                  },
-                }}
-              >
-                <CallEnd />
-              </IconButton>
-            </Tooltip>
-          </Box>
-        </Box>
-
-        {/* Settings Menu */}
-        <Menu
-          anchorEl={settingsAnchor}
-          open={Boolean(settingsAnchor)}
-          onClose={handleSettingsClose}
-          anchorOrigin={{
-            vertical: "top",
-            horizontal: "center",
-          }}
-          transformOrigin={{
-            vertical: "bottom",
-            horizontal: "center",
-          }}
-        >
-          <MenuItem
-            onClick={() => {
-              actions.setShowVideoTiles(!state.showVideoTiles);
-              handleSettingsClose();
-            }}
-          >
-            {state.showVideoTiles ? "Hide Video Tiles" : "Show Video Tiles"}
-          </MenuItem>
-          <MenuItem onClick={handleDeviceSettingsOpen}>
-            Voice & Video Settings
-          </MenuItem>
-          <Divider />
-          <MenuItem
-            onClick={() => {
-              navigate("/settings");
-              handleSettingsClose();
-            }}
-          >
-            All Settings
-          </MenuItem>
-        </Menu>
-
-        {/* Device Settings Dialog */}
-        <DeviceSettingsDialog
-          open={showDeviceSettings}
-          onClose={handleDeviceSettingsClose}
-          onDeviceChange={handleDeviceChange}
-        />
-
-        {/* Screen Source Picker Dialog */}
-        <ScreenSourcePicker
-          open={screenShare.showSourcePicker}
-          onClose={screenShare.handleSourcePickerClose}
-          onSelect={screenShare.handleSourceSelect}
-        />
-
-        {/* Capture Replay Dialog */}
-        <CaptureReplayModal
-          open={showCaptureModal}
-          onClose={() => setShowCaptureModal(false)}
-        />
-      </Paper>
-
-      {/* Debug Panel - Toggle with Ctrl+Shift+D */}
-      {showDebugPanel && <VoiceDebugPanel />}
-    </>
+    <Suspense fallback={null}>
+      <VoiceBottomBarContent />
+    </Suspense>
   );
 };
